pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "graph-to-agent"
version = "0.2.0"
description = "Visual agent orchestration framework for LLM workflows - Build multi-agent systems using graph-based visual programming"
readme = "README.md"
license = {text = "GPL-3.0"}
authors = [
    {name = "Kilian Lehn"}
]
maintainers = [
    {name = "Kilian Lehn"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
keywords = [
    "agents",
    "llm",
    "openai",
    "gpt",
    "graph",
    "orchestration",
    "multi-agent",
    "workflow",
    "ai",
    "machine-learning",
    "prompt-engineering",
    "agent-framework",
]
requires-python = ">=3.10"
dependencies = [
    "flask>=3.0.0",
    "werkzeug>=3.0.1",
    "openai>=1.12.0",
    "google-cloud-bigquery>=3.14.0",
    "google-auth>=2.27.0",
    "python-dotenv>=1.0.1",
    "requests>=2.31.0",
    "urllib3>=2.2.0",
    "cryptography>=42.0.0",
    "numpy>=1.26.0",
    "pandas>=2.2.0",
    "networkx>=3.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
    "pre-commit>=3.0",
    "detect-secrets>=1.4.0",
]
anthropic = [
    "anthropic>=0.20.0",
]
audio = [
    "openai-whisper>=20231117",
    "sounddevice>=0.4.0",
]
firebase = [
    "firebase-admin>=6.0.0",
    "firebase-functions>=0.1.0",
]
neo4j = [
    "neo4j>=5.0.0",
]
redis = [
    "redis>=5.0.0",
]
mem0 = [
    "mem0ai>=1.0.0",
]
hydradb = [
    "hydradb>=0.1.0",
]
memory = [
    "graph-to-agent[mem0,hydradb]",
]
graph = [
    "graph-to-agent[neo4j,redis]",
]
all = [
    "graph-to-agent[dev,anthropic,audio,firebase,neo4j,redis,mem0,hydradb]",
]

[project.urls]
Homepage = "https://github.com/kilian-lm/graph_to_agent"
Documentation = "https://github.com/kilian-lm/graph_to_agent#readme"
Repository = "https://github.com/kilian-lm/graph_to_agent"
Issues = "https://github.com/kilian-lm/graph_to_agent/issues"
Changelog = "https://github.com/kilian-lm/graph_to_agent/releases"

[project.scripts]
graph-to-agent = "app:app.run"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"graph_to_agent" = ["py.typed"]
"*" = ["templates/*.html", "static/**/*"]

[tool.ruff]
line-length = 120
target-version = "py310"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"legacy/*" = ["ALL"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=controllers --cov=logics --cov-report=term-missing"
python_files = ["test_*.py", "*_test.py"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
ignore_missing_imports = true
exclude = ["legacy/", "data_science/"]

[tool.coverage.run]
source = ["controllers", "logics", "logger"]
omit = ["*/tests/*", "legacy/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]
